+++ product/application/host/app.h
@@ -2,6 +2,7 @@
 #define _APP_H_
 #include "string.h"
 #include "bluenrg_sdk_api.h"
+#include "bluenrg_sdk_host_api.h"
 
 #endif //_APP_H_
